+++ src/components/Pagination/index.tsx
@@ -1,0 +1,23 @@
+import { Box, Button, HStack } from "@chakra-ui/react";
+import { PaginationItem } from "./PaginationItem";
+
+export function Pagination() {
+  return (
+    <HStack
+      mt="8"
+      justify="space-between"
+      align="center"
+      spacing="6"
+    >
+      <Box>
+        <strong>0</strong> - <strong>10</strong> de <strong>100</strong>
+      </Box>
+      <HStack>
+        <PaginationItem isCurrent number={1}/>
+        <PaginationItem number={2}/>
+        <PaginationItem number={3}/>
+        <PaginationItem number={4}/>
+      </HStack>
+    </HStack>
+  )
+}
