+++ src/components/Sidebar/index.tsx
@@ -1,5 +1,6 @@
 import { Box, Stack } from "@chakra-ui/react";
 import { RiContactsLine, RiDashboardLine, RiGitMergeLine, RiInputMethodLine } from "react-icons/ri";
+
 import { NavLink } from "./NavLink";
 import { NavSection } from "./NavSection";
